requirements.txt
fastapi
uvicorn
pandas
xgboost
joblib
scikit-learn
python-dotenv
requests
